# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: TF2 QuickServer API
  version: 1.0.0
  description: API for managing TF2 game servers. Designed for integration with the tf2pickup system.
servers:
  - url: https://tf2-quickserver.sonikro.com/api
    description: Production
  - url: /api
    description: Local / relative
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Auth0 M2M JWT token (Client Credentials flow). Use the azp claim as the client identifier.
    oauth2:
      type: oauth2
      description: Auth0 OAuth2 Client Credentials flow for M2M authentication.
      flows:
        clientCredentials:
          tokenUrl: https://tf2-quickserver.us.auth0.com/oauth/token
          scopes: {}
  schemas:
    Server:
      type: object
      properties:
        serverId:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        region:
          type: string
          example: us-east-1
        variant:
          type: string
          example: standard-competitive
        hostIp:
          type: string
          example: 1.2.3.4
        hostPort:
          type: integer
          example: 27015
        tvIp:
          type: string
          example: 1.2.3.4
        tvPort:
          type: integer
          example: 27020
        rconPassword:
          type: string
          example: rconpass123
        hostPassword:
          type: string
          example: serverpass123
        tvPassword:
          type: string
          example: tvpass123
        rconAddress:
          type: string
          example: 1.2.3.4:27015
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          description: Client ID of the creator
        status:
          type: string
          enum:
            - pending
            - ready
            - terminating
    CreateServerRequest:
      type: object
      required:
        - region
        - variantName
      properties:
        region:
          type: string
          example: us-east-1
          description: The AWS or OCI region to deploy in
        variantName:
          type: string
          example: standard-competitive
          description: The server variant to deploy
        extraEnvs:
          type: object
          additionalProperties:
            type: string
          example:
            STV_TITLE: My Pickup
            HOSTNAME: My TF2 Server
          description: Optional custom environment variables to pass to the server
    TaskAccepted:
      type: object
      properties:
        taskId:
          type: string
          example: task-1700000000000-abc123
    TaskStatus:
      type: object
      properties:
        taskId:
          type: string
          example: task-1700000000000-abc123
        type:
          type: string
          example: create-server-for-client
        status:
          type: string
          enum:
            - pending
            - running
            - completed
            - failed
        result:
          description: The Server object when status is completed
          $ref: '#/components/schemas/Server'
        error:
          type: string
          description: Error message when status is failed
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated client does not own this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
security:
  - bearerAuth: []
  - oauth2: []
x-auth0:
  domain: tf2-quickserver.us.auth0.com
  audience: https://tf2-quickserver.sonikro.com
paths:
  /api/servers:
    post:
      summary: Create a new TF2 server (async)
      description: |
        Enqueues a server creation task and returns a task ID immediately.
        Server creation typically takes 4-6 minutes. Poll `GET /api/tasks/{taskId}` to check status.
        When completed, the task result contains the full server details including credentials.
      tags:
        - Servers
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServerRequest'
      responses:
        '202':
          description: Server creation task accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      summary: List all servers for the authenticated client
      description: Returns all TF2 game servers associated with the authenticated API client.
      tags:
        - Servers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of servers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Server'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/servers/{serverId}:
    delete:
      summary: Delete a server (async)
      description: |
        Enqueues a server deletion task and returns a task ID immediately.
        Server deletion typically takes 1-3 minutes. Poll `GET /api/tasks/{taskId}` to check status.
        Only the client that created the server may delete it.
      tags:
        - Servers
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: serverId
          required: true
          schema:
            type: string
          description: The ID of the server to delete
      responses:
        '202':
          description: Server deletion task accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskAccepted'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/tasks/{taskId}:
    get:
      summary: Get the status of an async task
      description: |
        Poll this endpoint to check the status of a previously submitted async task
        (e.g. server creation or deletion). When `status` is `completed`, the `result`
        field contains the full server object. When `status` is `failed`, the `error`
        field contains the failure reason.
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
          description: The task ID returned by a create/delete server request
      responses:
        '200':
          description: Task status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
tags: []